requirements.txt
bottle==0.13.4
packaging==26.0
proxy_tools==0.1.0
PyQt6==6.10.2
PyQt6-Qt6==6.10.2
PyQt6-WebEngine==6.10.0
PyQt6-WebEngine-Qt6==6.10.2
PyQt6_sip==13.11.0
pywebview==6.1
QtPy==2.4.3
typing_extensions==4.15.0
